Create new module locations.rs Move code related to locations from holding_locations.rs
related to #3287
